#15 - Integrated folder generation with the GUI
