+++ FileManagement/Scripts/PageContent/index.css
@@ -16,4 +16,5 @@ body {
   font-family: monospace;
   width: 100%;
   height: 300px;
+  overflow-y: auto;
 }
