+++ FileManagement/Scripts/index.html
@@ -12,6 +12,7 @@
     <input class="btn" id="btnNewJamFolder" type="button" value="New Jam Folder" />
     <input class="btn" id="btnRemoveEmptyProjects" type="button" value="Remove Empty Projects" />
     <input class="btn" id="btnUpdateInventory" type="button" value="Update Inventory" />
+    <input class="btn" id="btnClearOutput" type="button" value="Clear Output" />
     
     <pre id="output"></pre>
