--- FileManagement/_new-jam-folder.cmd
@@ -1,9 +1,0 @@
-@ECHO OFF
-
-ECHO Generating new jam folder...
-
-node Scripts/tools/newprojectfolder.js --jam
-
-ECHO New jam folder is ready!
-
-PAUSE
