--- FileManagement/_new-project-folder.cmd
@@ -1,9 +1,0 @@
-@ECHO OFF
-
-ECHO Generating new project folder...
-
-node Scripts/tools/newprojectfolder.js
-
-ECHO New project folder is ready!
-
-PAUSE
