+++ FileManagement/setupWindows.cmd
@@ -4,5 +4,3 @@ ECHO Installing or updating Node.JS package...
 
 cd %~dp0\Scripts
 npm install
-
-ECHO Done!
